backend/requirements.txt
uvicorn>==0.27.1
fastapi>==0.109.2
docker>==7.0.0
pyaml>==23.12.0
pytest